refactor: extract methods from Drone to DroneTools.
